+++ scripts/install_redis.sh
@@ -1,20 +1,23 @@
 #!/usr/bin/env bash
 
 # Redis server installer
-# Min. Requirement  : GNU/Linux Ubuntu 16.04
-# Last Build        : 23/10/2019
+# Min. Requirement  : GNU/Linux Ubuntu 18.04
+# Last Build        : 11/12/2021
 # Author            : MasEDI.Net (me@masedi.net)
 # Since Version     : 1.0.0
 
 # Include helper functions.
-if [ "$(type -t run)" != "function" ]; then
-    BASEDIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )
+if [[ "$(type -t run)" != "function" ]]; then
+    BASE_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )
     # shellcheck disable=SC1091
-    . "${BASEDIR}/helper.sh"
+    . "${BASE_DIR}/helper.sh"
 fi
 
 # Make sure only root can run this installer script.
 requires_root
+
+# Make sure only supported distribution can run this installer script.
+preflight_system_check
 
 function add_redis_repo() {
     echo "Adding Redis repository..."
@@ -24,18 +27,18 @@ function add_redis_repo() {
 
     case ${DISTRIB_NAME} in
         debian)
-            if [ ! -f "/etc/apt/sources.list.d/dotdeb-stable.list" ]; then
+            if [[ ! -f "/etc/apt/sources.list.d/dotdeb-stable.list" ]]; then
                 run touch /etc/apt/sources.list.d/dotdeb-stable.list
                 run bash -c "echo -e 'deb http://ftp.utexas.edu/dotdeb/ stable all\ndeb-src http://ftp.utexas.edu/dotdeb/ stable all' > /etc/apt/sources.list.d/dotdeb-stable.list"
                 run bash -c "wget -qO - 'https://www.dotdeb.org/dotdeb.gpg' | apt-key add -"
-                run apt-get update -qq -y
+                run apt-get update -q -y
             else
                 info "Dotdeb repository already exists."
             fi
         ;;
         ubuntu)
             run add-apt-repository -y ppa:chris-lea/redis-server && \
-            run apt-get update -qq -y
+            run apt-get update -q -y
         ;;
         *)
             fail "Unable to add Redis, this GNU/Linux distribution is not supported."
@@ -47,12 +50,12 @@ function add_redis_repo() {
 function init_redis_install {
     local SELECTED_INSTALLER=""
 
-    if "${AUTO_INSTALL}"; then
-        if [[ -z "${REDIS_INSTALLER}" || "${REDIS_INSTALLER}" == "none" ]]; then
-            DO_INSTALL_REDIS="n"
-        else
+    if [[ "${AUTO_INSTALL}" == true ]]; then
+        if [[ "${INSTALL_REDIS}" == true ]]; then
             DO_INSTALL_REDIS="y"
             SELECTED_INSTALLER=${REDIS_INSTALLER:-"repo"}
+        else
+            DO_INSTALL_REDIS="n"
         fi
     else
         while [[ "${DO_INSTALL_REDIS}" != "y" && "${DO_INSTALL_REDIS}" != "n" ]]; do
@@ -60,7 +63,7 @@ function init_redis_install {
         done
     fi
 
-    if [[ ${DO_INSTALL_REDIS} == y* && "${INSTALL_REDIS}" == true ]]; then
+    if [[ ${DO_INSTALL_REDIS} == y* || ${DO_INSTALL_REDIS} == Y* ]]; then
         # Install menu.
         echo "Available Redis server installation method:"
         echo "  1). Install from Repository (repo)"
@@ -73,20 +76,16 @@ function init_redis_install {
         done
 
         case "${SELECTED_INSTALLER}" in
-            1|"repo")
+            1 | "repo")
                 # Add Redis repos.
                 add_redis_repo
 
                 echo "Installing Redis server from repository..."
 
                 # Install Redis.
-                if hash apt-get 2>/dev/null; then
-                    run apt-get install -qq -y redis-server redis-tools
-                else
-                    fail "Unable to install Redis, this GNU/Linux distribution is not supported."
-                fi
+                run apt-get install -q -y redis-server redis-tools
             ;;
-            2|"source")
+            2 | "source")
                 echo "Installing Redis server from source..."
                 
                 local CURRENT_DIR && \
@@ -108,18 +107,19 @@ function init_redis_install {
 
                     # Create Redis user. 
                     if [[ -z $(getent passwd redis) ]]; then
-                        if "${DRYRUN}"; then
-                            echo "Create redis user in dryrun mode."
-                        else
+                        if [[ "${DRYRUN}" != true ]]; then
                             run groupadd -r redis
                             run useradd -r -M -g redis redis
+                        else
+                            info "Create Redis user in dry run mode."
                         fi
                     fi
 
-                    if [ ! -d /etc/redis ]; then
+                    if [[ ! -d /etc/redis ]]; then
                         run mkdir /etc/redis
                     fi
-                    if [ ! -d /run/redis ]; then
+
+                    if [[ ! -d /run/redis ]]; then
                         run mkdir /run/redis
                         run chown redis:redis /run/redis
                         run chmod 770 /run/redis
@@ -131,18 +131,14 @@ function init_redis_install {
                 run cd "${CURRENT_DIR}" || error "Cannot change directory to ${CURRENT_DIR}"
             ;;
             *)
-                # Skip installation.
+                # Skip unsupported installation mode.
                 error "Installer method not supported. Redis installation skipped."
             ;;
         esac
 
         # Configure Redis.
-        if "${DRYRUN}"; then
-            info "Configuring Redis in dryrun mode."
-        else
-            if [ ! -f /etc/redis/redis.conf ]; then
-                run cp -f etc/redis/redis.conf /etc/redis/
-            fi
+        if [[ "${DRYRUN}" != true ]]; then
+            [[ ! -f /etc/redis/redis.conf ]] && run cp -f etc/redis/redis.conf /etc/redis/
 
             # Custom Redis configuration.
             local RAM_SIZE && \
@@ -171,8 +167,8 @@ maxmemory-policy allkeys-lru
 EOL
 
             # Is Redis password protected enable?
-            if "${REDIS_REQUIRE_PASSWORD}"; then
-                echo "Redis Requirepass is enabled..."
+            if [[ "${REDIS_REQUIRE_PASSWORD}" == true ]]; then
+                echo "Configure Redis requirepass password."
 
                 REDIS_PASSWORD=${REDIS_PASSWORD:-$(openssl rand -base64 64 | tr -dc 'a-zA-Z0-9' | fold -w 16 | head -n 1)}
 
@@ -180,10 +176,8 @@ EOL
                 cat >> /etc/redis/redis.conf <<EOL
 requirepass ${REDIS_PASSWORD}
 EOL
-                # Save config.
+                # Save data.
                 save_config "REDIS_PASSWORD=${REDIS_PASSWORD}"
-
-                # Save log.
                 save_log -e "Redis server requirepass is enabled, here is your authentication password: ${REDIS_PASSWORD}\nSave this password and use it to authenticate your Redis connection (typically use -a parameter)."
             fi
 
@@ -197,7 +191,7 @@ net.core.somaxconn=65535
 vm.overcommit_memory=1
 EOL
 
-            if [ ! -f /etc/rc.local ]; then
+            if [[ ! -f /etc/rc.local ]]; then
                 run touch /etc/rc.local
             fi
 
@@ -210,19 +204,21 @@ EOL
 sysctl -w net.core.somaxconn=65535
 echo never > /sys/kernel/mm/transparent_hugepage/enabled
 EOL
+        else
+            info "Redis configuration skipped in dry run mode."
         fi
 
         # Init Redis script.
-        if [ ! -f /etc/init.d/redis-server ]; then
+        if [[ ! -f /etc/init.d/redis-server ]]; then
             run cp -f etc/init.d/redis-server /etc/init.d/
             run chmod ugo+x /etc/init.d/redis-server
         fi
 
-        # Systemd Redis service.
-        if [ ! -f /lib/systemd/system/redis-server.service ]; then
+        # Setup Systemd service.
+        if [[ ! -f /lib/systemd/system/redis-server.service ]]; then
             run cp -f etc/systemd/redis-server.service /lib/systemd/system/
 
-            if [ ! -f /etc/systemd/system/redis.service ]; then
+            if [[ ! -f /etc/systemd/system/redis.service ]]; then
                 run ln -s /lib/systemd/system/redis-server.service /etc/systemd/system/redis.service
             fi
 
@@ -230,64 +226,36 @@ EOL
             run systemctl daemon-reload
         fi
 
-        # Restart Redis daemon.
+        # Restart and enable Redis on system boot.
         echo "Starting Redis server..."
-        if [ -f /etc/systemd/system/redis.service ]; then
-            run systemctl restart redis-server.service
 
-            # Enable Redis on system boot.
+        if [[ -f /etc/systemd/system/redis.service ]]; then
+            run systemctl restart redis-server.service
             run systemctl enable redis-server.service
         else
             run systemctl restart redis-server
         fi
 
-        if "${DRYRUN}"; then
-            info "Redis server installed in dryrun mode."
-        else
+        if [[ "${DRYRUN}" != true ]]; then
             if [[ $(pgrep -c redis-server) -gt 0 ]]; then
                 success "Redis server started successfully."
             else
                 info "Something went wrong with Redis installation."
             fi
+        else
+            info "Redis server started successfully in dry run mode."
         fi
-
-        # PHP version.
-        local PHPv="${1}"
-        if [ -z "${PHPv}" ]; then
-            PHPv=${PHP_VERSION:-"7.4"}
-        fi
-
-        # Install PHP Redis extension.
-        install_php_redis "$@"
     else
         info "Redis server installation skipped."
     fi
 }
 
-# Install PHP Redis extension.
-function install_php_redis() {
-    # PHP version.
-    local PHPv="${1}"
-    if [ -z "${PHPv}" ]; then
-        PHPv=${PHP_VERSION:-"7.4"}
-    fi
-
-    echo "Installing PHP ${PHPv} Redis extensions..."
-
-    if hash apt-get 2>/dev/null; then
-        run apt-get install -qq -y "php${PHPv}-redis"
-    else
-        fail "Unable to install PHP ${PHPv} Redis, this GNU/Linux distribution is not supported."
-    fi
-}
-
-
 echo "[Redis (Key-value) Server Installation]"
 
 # Start running things from a call at the end so if this script is executed
 # after a partial download it doesn't do anything.
-if [[ -n $(command -v redis-server) ]]; then
-    info "Redis key-value store server already exists. Installation skipped..."
+if [[ -n $(command -v redis-server) && "${FORCE_INSTALL}" != true ]]; then
+    info "Redis key-value store server already exists, installation skipped."
 else
     init_redis_install "$@"
 fi
